requirements.txt
uvicorn==0.21.0
fastapi==0.94.0
nebula3-python==3.4.0
starlette==0.26.0.post1